requirements.txt
bitcoin==1.1.42
blockchain==1.3.3
certifi==2017.4.17
chardet==3.0.4
ethereum-abi-utils==0.4.0
ethereum-utils==0.2.0
idna==2.5
Naked==0.1.31
pybitcointools==1.1.15
pycrypto==2.6.1
pylru==1.0.9
pysha3==1.0.2
PyYAML==3.12
requests==2.18.1
rlp==0.5.1
shellescape==3.4.1
termcolor==1.1.0
urllib3==1.21.1
web3==3.8.1
